+++ client/src/config/advancedSearch/inputTypes.ts
@@ -31,6 +31,26 @@ export const recordTypes: Record<string, Record<string, string>> = {
     person: 'Person',
     group: 'Group',
   },
+  item: {
+    HumanMadeObject: 'Physical Object',
+    DigitalObject: 'Digital Object',
+  },
+  work: {
+    VisualItem: 'Visual Item',
+    LinguisticObject: 'Textual Work',
+    Set: 'Set',
+  },
+  concept: {
+    Currency: 'Currency',
+    Language: 'Language',
+    Material: 'Material',
+    MeasurementUnit: 'Measurement Unit',
+    Type: 'General Concept',
+  },
+  event: {
+    Period: 'Period',
+    Activity: 'Activity',
+  },
 }
 
 export const comparators: Record<string, string> = {