+++ client/src/features/facets/Checkbox.tsx
@@ -10,7 +10,10 @@ import StyledCheckbox from '../../styles/features/facets/Checkbox'
 import ApiText from '../common/ApiText'
 import { getParamPrefix } from '../../lib/util/params'
 import { useAppDispatch, useAppSelector } from '../../app/hooks'
-import { addFacets, IFacetsSelected } from '../../redux/slices/facetsSlice'
+import {
+  addLastSelectedFacet,
+  IFacetsSelected,
+} from '../../redux/slices/facetsSlice'
 import { ResultsTab } from '../../types/ResultsTab'
 import { pushClientEvent } from '../../lib/pushClientEvent'
 
@@ -53,7 +56,7 @@ const Checkbox: React.FC<IProps> = ({
 
   // requires getting label before rendering it with capitalized content
   let label = ''
-  if (typeof facetValue === 'string') {
+  if (typeof facetValue === 'string' && !facetSection.includes('RecordType')) {
     const labelFromApi = ApiText(facetValue)
     label = labelFromApi !== null ? labelFromApi : ''
   } else {
@@ -77,7 +80,10 @@ const Checkbox: React.FC<IProps> = ({
       tab,
     )
     dispatch(
-      addFacets({ facetName: facetSection, facetUri: event.target.value }),
+      addLastSelectedFacet({
+        facetName: facetSection,
+        facetUri: event.target.value,
+      }),
     )
     pushClientEvent('Facets Checkbox', 'Unchecked', `Facet ${label}`)
     navigate(`${pathname}?${newSearchParams}`)
@@ -95,7 +101,9 @@ const Checkbox: React.FC<IProps> = ({
     params.set('q', JSON.stringify(criteria))
     params.set('facetRequest', 'true')
 
-    dispatch(addFacets({ facetName: facetSection, facetUri: strValue }))
+    dispatch(
+      addLastSelectedFacet({ facetName: facetSection, facetUri: strValue }),
+    )
     pushClientEvent('Facets Checkbox', 'Checked', `Facet ${label}`)
     navigate(`${pathname}?${params.toString()}`)
   }
@@ -125,6 +133,10 @@ const Checkbox: React.FC<IProps> = ({
 
       if (queryObj) {
         array.push(queryObj)
+      } else if (searchTermName === 'recordType') {
+        array.push({
+          [searchTermName as string]: value,
+        })
       } else if (idFacet && searchTermName) {
         array.push({
           [searchTermName as string]: {