+++ client/src/features/facets/SelectionContainer.tsx
@@ -108,7 +108,7 @@ const SelectionContainer: React.FC<IProps> = ({
             <StyledClearAllButton
               type="button"
               className="btn btn-link"
-              onClick={handleClearAll}
+              onClick={() => handleClearAll()}
             >
               Clear All
             </StyledClearAllButton>