+++ client/src/features/results/ObjectResults.tsx
@@ -66,7 +66,6 @@ const ObjectResults: React.FC<IProps> = ({ searchResponse }) => {
     <div className="row py-3">
       <FacetContainer
         facetsRequested={facetNamesLists.objects}
-        searchResponse={searchResponse}
         scope={searchScope.objects}
       />
       <Col xs={12} sm={12} md={12} lg={9}>